+++ collectfast/tests/command/test_copy_hooks.py
@@ -19,7 +19,7 @@ class BaseTestStrategy(Strategy[FileSystemStorage]):
 
     def __init__(self, remote_storage):
         super().__init__(remote_storage)
-        self.file_copied_hook = mock.MagicMock()
+        self.post_copy_hook = mock.MagicMock()
         self.file_skipped_hook = mock.MagicMock()
 
     def should_copy_file(
@@ -41,12 +41,12 @@ class FalseTestStrategy(BaseTestStrategy):
     COLLECTFAST_STRATEGY="collectfast.tests.command.test_copy_hooks.TrueTestStrategy",  # noqa: E501
     STATICFILES_STORAGE="django.core.files.storage.FileSystemStorage",
 )
-def test_calls_file_copied_hook_simple(case: TestCase) -> None:
+def test_calls_post_copy_hook_simple(case: TestCase) -> None:
     clean_static_dir()
     path = create_static_file()
     cmd = Command()
     cmd.run_from_argv(["manage.py", "collectstatic", "--noinput"])
-    cmd.strategy.file_copied_hook.assert_called_once_with(
+    cmd.strategy.post_copy_hook.assert_called_once_with(
         path.name,
         path.name,
         mock.ANY,
@@ -90,12 +90,12 @@ def test_calls_file_skipped_hook_collectfast(case: TestCase) -> None:
 
 @make_test_all_backends
 @live_test
-def test_call_file_copied_hook_all_backends(case: TestCase) -> None:
+def test_call_post_copy_hook_all_backends(case: TestCase) -> None:
     clean_static_dir()
     path = create_static_file()
     cmd = Command()
     cmd.run_from_argv(["manage.py", "collectstatic", "--noinput"])
-    cmd.strategy.file_copied_hook.assert_called_once_with(
+    cmd.strategy.post_copy_hook.assert_called_once_with(
         path.name,
         path.name,
         mock.ANY,